+++ CHANGELOG.md
@@ -11,7 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Breaking Changes
 
-This version of the dashboard requires Cumulus API >= v13.5.0 (TBD)
+This version of the dashboard requires Cumulus API >= v14.0.0
 
 ### Added