+++ CHANGELOG.md
@@ -1299,6 +1299,7 @@ Fix for serving the dashboard through the Cumulus API.
 ### Added
 
 - Expandable errors. [CUMULUS-394]
+- Added d3-dag dependency to replace dagre-d3. [CUMULUS-3107]
 
 ### Changed
 
@@ -1307,6 +1308,7 @@ Fix for serving the dashboard through the Cumulus API.
 - Tests use ava instead of tape. [CUMULUS-418]
 - Remove `defaultVersion` from the config. To use a particular version of the
   API, just set that in the API URL.
+- Upated D3 dependency to the latest version 7.8.2. [CUMULUS-3107]
 
 ## v1.0.1 - 2018-03-07